Extract TermWithDefinition to a reusable module
